+++ pro/ai/mcp/client/src/test/kotlin/org/http4k/ai/mcp/client/http/HttpStreamingMcpClientTest.kt
@@ -27,11 +27,13 @@ import org.http4k.ai.mcp.protocol.ClientCapabilities
 import org.http4k.ai.mcp.protocol.ServerMetaData
 import org.http4k.ai.mcp.protocol.Version
 import org.http4k.ai.mcp.server.capability.ServerCompletions
+import org.http4k.ai.mcp.server.capability.ServerInitializer
 import org.http4k.ai.mcp.server.capability.ServerPrompts
 import org.http4k.ai.mcp.server.capability.ServerResources
 import org.http4k.ai.mcp.server.capability.ServerTools
 import org.http4k.ai.mcp.server.http.HttpSessions
 import org.http4k.ai.mcp.server.http.HttpStreamingMcp
+import org.http4k.ai.mcp.server.capability.SimpleInitializeHandler
 import org.http4k.ai.mcp.server.protocol.McpProtocol
 import org.http4k.ai.mcp.server.security.OAuthMcpSecurity
 import org.http4k.ai.mcp.server.sessions.SessionProvider
@@ -127,10 +129,8 @@ class HttpStreamingMcpClientTest : McpStreamingClientContract<Sse>() {
         })
 
         val protocol = McpProtocol(
-            ServerMetaData(McpEntity.of("David"), Version.of("0.0.1")),
-            HttpSessions(
-                sessionProvider = SessionProvider.Random(Random(0)),
-            ).apply { start() },
+            HttpSessions(sessionProvider = SessionProvider.Random(Random(0))).apply { start() },
+            ServerInitializer(SimpleInitializeHandler(ServerMetaData(McpEntity.of("David"), Version.of("0.0.1")))),
             tools,
             ServerResources(
                 Resource.Static(