+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API
 changes with their rationale when appropriate:
 
-### v6.40.2.0
+### v6.41.0.0
 - **http4k-***: Upgrade versions
-- **http4k-tools-verify**: Support for key rotation.
+- **http4k-ai-mcp-sdk**: Add support for custom MCP session handshaking by introducing `InitializeHandler`.
+- **http4k-tools-verify**: Support for key rotation when signing cosigned artifacts.
 
 ### v6.40.1.0
 - **http4k-***: Upgrade versions